Support linux + clang

+++ imoguizmo.hpp
@@ -25,6 +25,7 @@ SOFTWARE.
 #pragma once
 #include <imgui.h>
 #include <cmath>
+#include <cstring>
 #include <vector>
 #include <algorithm>
 
@@ -89,7 +90,7 @@ namespace ImOGuizmo {
 			const auto lineEndPositive = ImVec2{ center.x + axis.x, center.y + axis.y };
 			internal::config.mDrawList->AddLine(center, lineEndPositive, color, thickness);
 			internal::config.mDrawList->AddCircleFilled(lineEndPositive, radius, color);
-			const auto textPosX = ImVec2{ floor(lineEndPositive.x - 3.0f), floor(lineEndPositive.y - 6.0f) };
+			const auto textPosX = ImVec2{ static_cast<float>(floor(lineEndPositive.x - 3.0f)), static_cast<float>(floor(lineEndPositive.y - 6.0f)) };
 			if (selected) {
 				internal::config.mDrawList->AddCircle(lineEndPositive, radius, IM_COL32_WHITE, 0, 1.1f);
 				internal::config.mDrawList->AddText(textPosX, IM_COL32_WHITE, text);
